requirements.txt
tensorflow
scikit-learn
numpy
pandas
matplotlib
seaborn
fastapi
uvicorn
pydantic
ipykernel
simpy
supabase
python-dotenv